+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 - Add JSON encoder to handle numpy parameters for remote serialization
+- Remote robots now support grippers and cameras
 
 ## [0.8.1] - 2026-02-09
 - Parse mass property for geom tags when importing MuJoCo XML files